requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6

xgboost==2.0.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
shap==0.44.1

# For sentiment analysis
nltk==3.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# LLM (free)
groq>=0.9.0

# RAG
sentence-transformers>=2.2.0
pinecone>=3.0.0
pypdf2>=3.0.0

# Auth
PyJWT>=2.8.0
email-validator>=2.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

